Added basic files for proper testing/functions
Added new function - RegisterUserPage and RegisterUserTest. It works quite simply, the automated test fills the form with the data we set in TestData. Once filled, the test clicks the Continue button and then compares the link we got after signing up with the link we should get after signing up.